Player/MarioWait: waitingCommonEvents (u8)canSquat() narrow cast (88.6->89.1%)
Target narrows canSquat() result with clrlwi. (8-bit) but ours emitted cmpwi
(wide). (u8) cast forces the narrow. Residual = separate 40B frame inflation.

+++ src/Player/MarioWait.cpp
@@ -168,7 +168,7 @@ BOOL TMario::waitingCommonEvents()
 		return 1;
 	}
 
-	if (canSquat()) {
+	if ((u8)canSquat()) {
 		mForwardVel = 0.0f;
 		changePlayerStatus(0x0C018220, 0, false);
 		return 1;
